Create handler for MarketInsufficientSubsidy
