+++ chain/prediction-markets.ts
@@ -48,4 +48,36 @@ export namespace PredictionMarkets {
       return valid;
     }
   }
+
+  /**
+   * A market was discarded after failing to gather enough subsidy. \[market_id\]
+   *
+   *  Event parameters: [MarketIdOf<T>, ]
+   */
+  export class MarketInsufficientSubsidyEvent {
+    public readonly expectedParamTypes = ["MarketIdOf<T>"];
+
+    constructor(public readonly ctx: SubstrateEvent) {}
+
+    get params(): [MarketIdOf] {
+      return [
+        createTypeUnsafe<MarketIdOf & Codec>(typeRegistry, "MarketIdOf", [
+          this.ctx.params[0].value,
+        ]),
+      ];
+    }
+
+    validateParams(): boolean {
+      if (this.expectedParamTypes.length !== this.ctx.params.length) {
+        return false;
+      }
+      let valid = true;
+      this.expectedParamTypes.forEach((type, i) => {
+        if (type !== this.ctx.params[i].type) {
+          valid = false;
+        }
+      });
+      return valid;
+    }
+  }
 }
